+++ lua/pyworks/health.lua
@@ -197,4 +197,3 @@ function M.check()
 end
 
 return M
-