+++ lua/pyworks/languages/julia.lua
@@ -316,4 +316,3 @@ function M.setup_repl()
 end
 
 return M
-